requirements.txt
# Vector DB
chromadb

# Core Langchain Stack (pinned for compatibility)
openai
langchain
langchain-community
langchain-openai
python-dotenv
numpy

# SQL & DB
sqlglot>=18.0.0
sqlite-utils>=3.35

# Frontend (Streamlit)
streamlit>=1.25.0

# Testing & Dev
pytest
httpx
requests
